requirements.txt
Django==4.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
python-decouple==3.8
PyJWT==2.8.0
bcrypt==4.1.2
djangorestframework-simplejwt==5.3.0 